requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-spectacular==0.18.1
inflection==0.5.1
jsonschema==3.2.0
Pillow==8.3.2
PyJWT==2.1.0
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
